Feature/MSB-4132

Approved-by: Ronald Wentworth

+++ src/main/java/com/moblize/ms/dailyops/dto/Section.java
@@ -26,14 +26,14 @@ import lombok.ToString;
 public class Section {
 
     @JsonProperty("a")
-    public Integer all;
+    public Double all;
     @JsonProperty("s")
-    public Integer surface;
+    public Double surface;
     @JsonProperty("i")
-    public Integer intermediate;
+    public Double intermediate;
     @JsonProperty("c")
-    public Integer curve;
+    public Double curve;
     @JsonProperty("l")
-    public Integer lateral;
+    public Double lateral;
 
 }
